Extract TransferDetailsModal
