+++ packages/ui/src/components/TransferModal.tsx
@@ -3,13 +3,10 @@ import styled from 'styled-components'
 import BN from 'bn.js'
 import { BorderRad, Colors } from '../constants'
 import { useBalances } from '../hooks/useBalances'
-import { useNumberInput } from '../hooks/useNumberInput'
 import { Account } from '../hooks/types'
-import { Modal, ModalBody, ModalFooter, ModalHeader } from './modal'
-import { ButtonPrimaryMedium, ButtonSecondarySmall } from './buttons/Buttons'
-import { AccountInfo } from './AccountInfo'
-import { TokenValue } from './TokenValue'
+import { Modal, ModalBody, ModalHeader } from './modal'
 import { SignTransferModal } from './transfer/SignTransferModal'
+import { TransferDetailsModal } from './transfer/TransferDetailsModal'
 
 interface Props {
   onClose: () => void
@@ -20,76 +17,26 @@ interface Props {
 type ModalState = 'SEND_TOKENS' | 'SIGN_TRANSACTION' | 'SENDING'
 
 export function TransferModal({ from, to, onClose }: Props) {
-  const balances = useBalances([from, to])
-  const [amount, setAmount] = useNumberInput(0)
+  const { hasBalances } = useBalances([from, to])
   const [step, setStep] = useState<ModalState>('SEND_TOKENS')
+  const [amount, setAmount] = useState<BN>(new BN(0))
 
-  const isZero = new BN(amount).lte(new BN(0))
+  const onAccept = (amount: BN) => {
+    setAmount(amount)
+    setStep('SIGN_TRANSACTION')
+  }
 
-  if (!balances.hasBalances) {
+  if (!hasBalances) {
     return (
       <Modal>
+        <ModalHeader onClick={onClose} title={''} />
         <ModalBody>Loading balances...</ModalBody>
       </Modal>
     )
   }
 
-  const transferableBalance = balances?.map[from.address]?.total
-
-  const isOverBalance = new BN(amount).gt(transferableBalance)
-  const isTransferDisabled = isZero || isOverBalance
-  const title = step === 'SIGN_TRANSACTION' ? 'Authorize transaction' : 'Send tokens'
-
-  const setHalf = () => setAmount(transferableBalance.div(new BN(2)).toString())
-  const setMax = () => setAmount(transferableBalance.toString())
-
   if (step === 'SEND_TOKENS') {
-    return (
-      <Modal>
-        <ModalHeader onClick={onClose} title={title} />
-        <ModalBody>
-          <Row>
-            <FormLabel>From</FormLabel>
-            <LockedAccount>
-              <AccountInfo account={from} />
-              <TransactionInfoRow>
-                <InfoTitle>Transferable balance</InfoTitle>
-                <InfoValue>
-                  <TokenValue value={transferableBalance} />
-                </InfoValue>
-              </TransactionInfoRow>
-            </LockedAccount>
-          </Row>
-          <TransactionAmount>
-            <AmountInputBlock>
-              <AmountInputLabel>Number of tokens</AmountInputLabel>
-              <AmountInput value={amount} onChange={(event) => setAmount(event.target.value)} placeholder={'0'} />
-            </AmountInputBlock>
-            <AmountButtons>
-              <AmountButton onClick={setHalf}>Use half</AmountButton>
-              <AmountButton onClick={setMax}>Use max</AmountButton>
-            </AmountButtons>
-          </TransactionAmount>
-          <Row>
-            <FormLabel>Destination account</FormLabel>
-            <AccountRow>
-              <AccountInfo account={to} />
-              <TransactionInfoRow>
-                <InfoTitle>Total balance</InfoTitle>
-                <InfoValue>
-                  <TokenValue value={balances?.map[to.address]?.total} />
-                </InfoValue>
-              </TransactionInfoRow>
-            </AccountRow>
-          </Row>
-        </ModalBody>
-        <ModalFooter>
-          <ButtonPrimaryMedium onClick={() => setStep('SIGN_TRANSACTION')} disabled={isTransferDisabled}>
-            Transfer tokens
-          </ButtonPrimaryMedium>
-        </ModalFooter>
-      </Modal>
-    )
+    return <TransferDetailsModal onClose={onClose} from={from} to={to} onAccept={onAccept} />
   }
 
   return <SignTransferModal onClose={onClose} from={from} amount={amount} to={to} />
@@ -132,35 +79,6 @@ export const AmountInputBlock = styled.div`
   display: flex;
   flex-direction: column;
 `
-const AmountInputLabel = styled.span`
-  margin-bottom: 8px;
-  font-size: 14px;
-  line-height: 20px;
-  font-weight: 700;
-  color: ${Colors.Black[900]};
-`
-const AmountInput = styled.input`
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-  width: 100%;
-  padding: 12px 16px;
-  border: 1px solid ${Colors.Black[900]};
-  border-radius: ${BorderRad.s};
-  font-size: 14px;
-  line-height: 20px;
-  font-weight: 700;
-  text-align: right;
-`
-const AmountButtons = styled.div`
-  display: inline-grid;
-  grid-auto-flow: column;
-  grid-column-gap: 8px;
-  width: fit-content;
-  height: 46px;
-  align-items: center;
-`
-const AmountButton = styled(ButtonSecondarySmall)``
 
 export const TransactionInfo = styled.div`
   display: flex;
